Support path mapping, x-plat
The original script only supported Windows paths ('\' path separators), and
the script had to be run from the computer hosting Plex (or at least have the
exact same path mappings). This change allows for either '\' or '/' path
separators, and also lets the user define path mappings if the script is being
run on a different machine, but Plex media is still available under a different
relative path.